Fix usec->nsec and auto->log in timer and else

+++ src/tools/Timer.h
@@ -11,7 +11,7 @@ class Timer {
 public:
     Timer();
     void restart();
-    long long getElapsedNanosec() const;
+    long getElapsedNanosec() const;
     void sleepNanosec(long ns) const;
 
 private:
